requirements.txt
absl-py==0.8.0
astor==0.8.0
astroid==2.3.1
attrs==19.2.0
backcall==0.1.0
bleach==3.1.0
certifi==2020.4.5.2
click==7.1.2
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
Flask==1.1.2
google-pasta==0.1.7
grpcio==1.24.1
h5py==2.10.0
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.3
json5==0.8.5
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.3
jupyter-console==6.0.0
jupyter-core==4.5.0
jupyterlab==1.1.4
jupyterlab-server==1.0.6
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.2
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.6.0
nbformat==4.4.0
notebook==6.0.1
numpy==1.17.2
opencv-python==4.2.0.34
pandocfilters==1.4.2
parso==0.5.1
pickleshare==0.7.5
Pillow==7.1.2
prometheus-client==0.7.1
prompt-toolkit==2.0.10
ptyprocess==0.6.0
Pygments==2.4.2
pylint==2.4.2
pyparsing==2.4.2
pyrsistent==0.15.4
pyzmq==18.1.0
qtconsole==4.5.5
Send2Trash==1.5.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
traitlets==4.3.3
typed-ast==1.4.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wrapt==1.11.2
